map section added, div box broken
